+++ CHANGELOG.md
@@ -15,6 +15,8 @@ https://github.com/nwnxee/unified/compare/build8193.16...HEAD
 - Events: added Spell Interruption events to SpellEvents
 - Events: added skippable Has Feat event to FeatEvents
 - Events: added Journal Open/Close events to JournalEvents
+- Events: added Detect Mode event to StealthEvents
+- Events: added Combat Enter/Exit events to CombatEvents
 - Events: Stealth Mode can now bypass or perform Hide in Plain Sight with return values of "0" or "1" respectively
 - Tweaks: `NWNX_TWEAKS_HIDE_PLAYERS_ON_CHAR_LIST`
 - Tweaks: `NWNX_TWEAKS_FIX_ARMOR_DEX_BONUS_UNDER_ONE`